refactor(telemetry): rename verifiers package to prime-intellect
Use latitude-telemetry-prime-intellect / packages/telemetry/prime-intellect
so the Prime Intellect integration is obvious; Verifiers stays in docs and
span metadata as the framework name.
