+++ docs/docs.json
@@ -159,7 +159,7 @@
               "telemetry/hermes",
               "telemetry/openclaw",
               "telemetry/pi-coding-agent",
-              "telemetry/verifiers"
+              "telemetry/prime-intellect"
             ]
           }
         ]
