+++ packages/telemetry/prime-intellect/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "latitude-telemetry-verifiers"
+name = "latitude-telemetry-prime-intellect"
 version = "0.1.0"
 description = "Export Prime Intellect Verifiers eval rollouts to Latitude as OTLP traces and scores"
 authors = [{ name = "Latitude Data SL", email = "hello@latitude.so" }]
@@ -7,24 +7,24 @@ maintainers = [{ name = "Latitude Data SL", email = "hello@latitude.so" }]
 readme = "README.md"
 license = "MIT"
 keywords = [
-    "verifiers",
     "prime-intellect",
+    "verifiers",
     "telemetry",
     "otlp",
     "opentelemetry",
     "latitude",
     "evals",
 ]
-urls.repository = "https://github.com/latitude-dev/latitude-llm/tree/main/packages/telemetry/verifiers"
-urls.homepage = "https://github.com/latitude-dev/latitude-llm/tree/main/packages/telemetry/verifiers#readme"
-urls.documentation = "https://github.com/latitude-dev/latitude-llm/tree/main/packages/telemetry/verifiers#readme"
+urls.repository = "https://github.com/latitude-dev/latitude-llm/tree/main/packages/telemetry/prime-intellect"
+urls.homepage = "https://github.com/latitude-dev/latitude-llm/tree/main/packages/telemetry/prime-intellect#readme"
+urls.documentation = "https://github.com/latitude-dev/latitude-llm/tree/main/packages/telemetry/prime-intellect#readme"
 requires-python = ">=3.11, <3.15"
 dependencies = [
     "certifi>=2024.2.2",
 ]
 
 [project.scripts]
-latitude-verifiers-export = "latitude_telemetry_verifiers.__main__:main"
+latitude-prime-intellect-export = "latitude_telemetry_prime_intellect.__main__:main"
 
 [dependency-groups]
 dev = [
@@ -56,4 +56,4 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/latitude_telemetry_verifiers"]
+packages = ["src/latitude_telemetry_prime_intellect"]
