+++ packages/telemetry/prime-intellect/src/latitude_telemetry_prime_intellect/config.py
@@ -10,7 +10,7 @@ Optional env vars:
   LATITUDE_NO_CONTENT     - "true" to export structure/timing without prompts/tool I/O
   LATITUDE_EXPORT_SCORES  - "true" (default) to POST reward/metrics as custom scores
   LATITUDE_DEBUG          - "true" for verbose logging
-  LATITUDE_VERIFIERS_TELEMETRY_ENABLED / LATITUDE_TELEMETRY_ENABLED - master switch
+  LATITUDE_PRIME_INTELLECT_TELEMETRY_ENABLED / LATITUDE_TELEMETRY_ENABLED - master switch
 """
 
 from __future__ import annotations
@@ -35,7 +35,7 @@ _SSL_CONTEXT = _ssl_context()
 
 logger = logging.getLogger(__name__)
 
-SCOPE_NAME = "latitude-telemetry-verifiers"
+SCOPE_NAME = "latitude-telemetry-prime-intellect"
 PKG_VERSION = "0.1.0"
 
 
@@ -52,9 +52,9 @@ def _load_config() -> Dict[str, Any]:
     project = _env("LATITUDE_PROJECT") or _env("LATITUDE_PROJECT_SLUG")
     base_url = _env("LATITUDE_BASE_URL") or "https://ingest.latitude.so"
     api_base_url = _env("LATITUDE_API_BASE_URL") or "https://api.latitude.so"
-    enabled_flag = _env("LATITUDE_VERIFIERS_TELEMETRY_ENABLED") or _env("LATITUDE_TELEMETRY_ENABLED")
+    enabled_flag = _env("LATITUDE_PRIME_INTELLECT_TELEMETRY_ENABLED") or _env("LATITUDE_TELEMETRY_ENABLED")
     enabled = enabled_flag.lower() not in {"0", "false", "no"} if enabled_flag else True
-    no_content = _env("LATITUDE_VERIFIERS_NO_CONTENT") or _env("LATITUDE_NO_CONTENT")
+    no_content = _env("LATITUDE_PRIME_INTELLECT_NO_CONTENT") or _env("LATITUDE_NO_CONTENT")
     allow_content = no_content.lower() not in {"1", "true", "yes"} if no_content else True
     export_scores_flag = _env("LATITUDE_EXPORT_SCORES")
     export_scores = export_scores_flag.lower() not in {"0", "false", "no"} if export_scores_flag else True
@@ -87,4 +87,4 @@ def _reset_config_for_tests() -> None:
 
 def _debug(message: str) -> None:
     if _config().get("debug"):
-        logger.info("Latitude verifiers telemetry: %s", message)
+        logger.info("Latitude Prime Intellect telemetry: %s", message)
