+++ pnpm-workspace.yaml
@@ -3,7 +3,7 @@ packages:
   - packages/**
   - "!packages/telemetry/python/**"
   - "!packages/telemetry/hermes/**"
-  - "!packages/telemetry/verifiers/**"
+  - "!packages/telemetry/prime-intellect/**"
   - "!packages/platform/op-gepa/python/**"
   - "!packages/telemetry/typescript/examples/flue-app"
   - "!packages/telemetry/typescript/examples/flue-app/**"
