+++ specs/telemetry-qa.md
@@ -367,7 +367,7 @@ Run via the uv venv at `packages/telemetry/python/.venv`.
 | 55 | OpenClaw CLI installer | `packages/telemetry/openclaw-cli` | One-shot installer for the OpenClaw plugin | ⚠️ | ⬜ | Installer/config tool — verify wiring, not trace shape directly. |
 | 56 | Pi agent telemetry | `packages/telemetry/pi` | Pi coding-agent extension → streams sessions as OTLP traces | ⚠️ | ⬜ | Needs the Pi agent installed. |
 | 59 | Hermes telemetry | `packages/telemetry/hermes` | **pip** plugin for Nous Research's **Hermes Agent** harness → streams sessions as OTLP traces (user prompts, model turns, tool calls/results, token usage, real system prompt). Same family as claude-code/pi/openclaw. | ⚠️ | ⬜ | **NEW (merged to development, picked up via the 2026-06-26 rebase; PRs incl. #3681).** Python pkg `latitude-telemetry-hermes` v0.1.0, entry point `hermes_agent.plugins` → `latitude`. Needs Hermes Agent installed in the **same** venv (official installer uses `~/.hermes/hermes-agent/venv`); enable via `~/.hermes/config.yaml` (**not** `hermes plugins enable` — hermes-agent#23802 reports pip plugins as "not installed" to the CLI). Has `tests/test_plugin.py`. Verify OTLP trace shape like the other harness integrations when reached. |
-| 61 | Verifiers (Prime Intellect) telemetry | `packages/telemetry/verifiers` | **pip** library/CLI for Prime Intellect **Verifiers** eval rollouts → OTLP traces + optional custom scores from rewards/metrics. Same family as hermes/claude-code/pi. | ⚠️ | ⬜ | Python pkg `latitude-telemetry-verifiers` v0.1.0. No host plugin entry point — call `export_episodes` / `make_on_complete` from eval scripts, or `latitude-verifiers-export export <results_dir>` post-hoc. Docs: `docs/telemetry/verifiers.md`. Has `tests/test_mapper.py` + `tests/test_export.py`. |
+| 61 | Prime Intellect (Verifiers) telemetry | `packages/telemetry/prime-intellect` | **pip** library/CLI for Prime Intellect **Verifiers** eval rollouts → OTLP traces + optional custom scores from rewards/metrics. Same family as hermes/claude-code/pi. | ⚠️ | ⬜ | Python pkg `latitude-telemetry-prime-intellect` v0.1.0. No host plugin entry point — call `export_episodes` / `make_on_complete` from eval scripts, or `latitude-prime-intellect-export export <results_dir>` post-hoc. Docs: `docs/telemetry/prime-intellect.md`. Has `tests/test_mapper.py` + `tests/test_export.py`. |
 | 60 | ElevenLabs Agents | `docs/telemetry/frameworks/elevenlabs.mdx` (docs-only) | Hosted voice-agent platform; observed via ElevenLabs **Custom LLM** → point the agent at a self-run OpenAI-compatible proxy instrumented with Latitude. | ⚠️ | ✅ | **DONE (review-only).** Not a Latitude package — the actual telemetry is the **OpenAI instrumentation** (#1) running on your proxy, so span parsing is already covered. Docs `elevenlabs.mdx` reviewed = good + nav-registered. **Icon:** custom `elevenlabs.tsx` (theme-aware bars). **In-app instructions ADDED:** TS+Py, custom-LLM proxy snippets (express / fastapi) with `instrumentations:{openai}`, packages `openai express` / `openai fastapi uvicorn`. |
 
 > **Eve/Flue/ElevenLabs polish (this PR, 2026-06-26):** added provider icons as `.tsx` (`elevenlabs.tsx`, `flue.tsx`; Eve reuses `VercelIcon`), wired into `provider-map.ts` + barrel; added **in-app onboarding instructions** for all three to `onboarding-integration-snippets.ts` + `telemetry-instructions.tsx`. Docs pages + examples reviewed (no code bugs). Remaining non-SDK integrations (#54 OpenClaw paths, #55, #56 Pi, #59 Hermes) deferred to a follow-up PR.
